Lambda/Docker/ingestion_requirements.txt
boto3==1.34.116
langchain==0.2.7
langchain-core==0.2.12
pytesseract==0.3.10
Pillow==11.0.0
numpy==1.26.4
PyMuPDF==1.24.9
pdf2image==1.17.0
